requirements.txt
fpdf == 1.7.2
PyQt5 == 5.15.6
PyQt5-stubs == 5.15.2.0
pyserial == 3.5
pyqtgraph == 0.12.3